# openapi.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8085","description":"Generated server url"}],"paths":{"/api/v1/webhooks/vyapar":{"post":{"tags":["webhook-controller"],"operationId":"handleVyaparWebhook","parameters":[{"name":"X-VyaparGateway-Signature","in":"header","required":false,"schema":{"type":"string"}},{"name":"X-VyaparGateway-Event-Id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/webhooks/razorpay":{"post":{"tags":["webhook-controller"],"operationId":"handleRazorpayWebhook","parameters":[{"name":"x-razorpay-signature","in":"header","required":false,"schema":{"type":"string"}},{"name":"x-razorpay-event-id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/webhooks/cashfree":{"post":{"tags":["webhook-controller"],"operationId":"handleCashfreeWebhook","parameters":[{"name":"x-webhook-signature","in":"header","required":false,"schema":{"type":"string"}},{"name":"x-webhook-timestamp","in":"header","required":false,"schema":{"type":"string"}},{"name":"x-webhook-event-id","in":"header","required":false,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/payments/refund":{"post":{"tags":["payment-controller"],"operationId":"refundOrder","parameters":[{"name":"gateway","in":"query","required":true,"schema":{"type":"string","enum":["RAZORPAY","CASHFREE","VYAPAR"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RefundRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/payments/create-order":{"post":{"tags":["payment-controller"],"operationId":"createOrder","parameters":[{"name":"gateway","in":"query","required":true,"schema":{"type":"string","enum":["RAZORPAY","CASHFREE","VYAPAR"]}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateOrderRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/internal/admin/payments/dlq/retry":{"post":{"tags":["admin-dlq-controller"],"operationId":"retryDlqEvent","parameters":[{"name":"topic","in":"query","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ApiResponseString"}}}}}}}},"components":{"schemas":{"RefundRequest":{"required":["amountInInr","gatewayOrderId"],"type":"object","properties":{"gatewayOrderId":{"type":"string"},"amountInInr":{"type":"number"},"reason":{"type":"string"}}},"CreateOrderRequest":{"required":["amountInInr","internalOrderId"],"type":"object","properties":{"internalOrderId":{"pattern":"^([a-zA-Z0-9_]+_)?([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12})$","type":"string"},"amountInInr":{"type":"number"},"customerPhone":{"type":"string"}}},"ApiResponseString":{"type":"object","properties":{"success":{"type":"boolean"},"message":{"type":"string"},"data":{"type":"string"},"timestamp":{"type":"string","format":"date-time"}}}}}}